テストコードの分離
