+++ lib/fizz_buzz.rb
@@ -12,12 +12,3 @@ def fizz_buzz(n)
 end
 
 
-require 'minitest/autorun'
-
-class FizzBuzzTest < Minitest::Test
-  def test_fizz_buzz
-    assert_equal '1',fizz_buzz(1)
-    assert_equal '2',fizz_buzz(2)
-    assert_equal 'Fizz',fizz_buzz(3)
-  end
-end
